Fix image_persistence
