+++ src/openag_brain/utils.py
@@ -1,5 +1,6 @@
 import os
 import rospkg
+from re import match
 
 def resolve_fixtures(fixtures):
     """
@@ -11,3 +12,20 @@ def resolve_fixtures(fixtures):
     fixtures_path = os.path.join(pkg_path, "fixtures")
     resolved = [os.path.join(fixtures_path, fixture_name + ".json") for fixture_name in fixtures]
     return resolved
+
+def read_environment_from_ns(namespace):
+    """
+    Given a ROS topic name, attempt to match an environment ID in topic
+    namespace.
+
+    Use:
+
+        read_environment_from_ns(rospy.get_namespace())
+    """
+    result = match("/environments/(\w+)/", namespace)
+    if not result:
+        raise ValueError(
+            "No environment id found in namespace \"{}\".".format(namespace)
+        )
+    environment_id = result.group(1)
+    return environment_id
